Rudr

+++ server/FrcsPos/Response/DTO/SuperAdminDashboardDTO.cs
@@ -24,7 +24,7 @@ namespace FrcsPos.Response.DTO
         public int TotalUsers { get; set; }
         public int TotalProducts { get; set; }
         public long TotalMedia { get; set; }
-        public int TotalSales { get; set; }
+        public decimal TotalSales { get; set; }
 
         public List<NotificationDTO> Notifications { get; set; } = new();
